added texting

+++ db/migrate/20190402174125_create_users.rb
@@ -4,7 +4,7 @@ class CreateUsers < ActiveRecord::Migration[5.2]
       t.string :name
       t.string :username
       t.string :email
-      t.integer :number
+      t.float :number
       t.string :password_digest
       t.string :bio
       t.string :avatar, default: "https://cdn0.iconfinder.com/data/icons/elasto-online-store/26/00-ELASTOFONT-STORE-READY_user-circle-512.png"
